+++ flink-doris-connector/src/main/java/org/apache/doris/flink/sink/writer/serializer/RecordWithMetaSerializer.java
@@ -1,4 +1,21 @@
-package org.apache.doris.flink.sink.writer;
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package org.apache.doris.flink.sink.writer.serializer;
 
 import org.apache.commons.lang3.StringUtils;
 import org.apache.doris.flink.sink.batch.RecordWithMeta;
@@ -9,7 +26,7 @@ import org.slf4j.LoggerFactory;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 
-public class RecordWithMetaSerializer implements DorisRecordSerializer<RecordWithMeta>{
+public class RecordWithMetaSerializer implements DorisRecordSerializer<RecordWithMeta> {
     private static final Logger LOG = LoggerFactory.getLogger(RecordWithMetaSerializer.class);
 
     @Override
